+++ libs/ng-diagrams/src/index.ts
@@ -1,0 +1,4 @@
+export * from './lib/hierarchical-box-diagram/box';
+export * from './lib/hierarchical-box-diagram/hierarchical-box-diagram.component';
+
+export * from './lib/ng-diagrams.module';