+++ libs/ng-diagrams/src/lib/hierarchical-box-diagram/box-node.ts
@@ -1,0 +1,7 @@
+import { Box } from './box';
+
+export interface BoxNode {
+  data: Box;
+  children?: BoxNode[];
+  parent?: BoxNode;
+}