+++ libs/ui/src/lib/form-item/form-item.component.stories.ts
@@ -6,10 +6,45 @@ export default {
   title: 'Angular/Form/Form item',
   component: FormItemComponent,
   parameters: { notes },
+  argTypes: {
+    displayMessage: {
+      name: 'Show form message?',
+      table: {
+        category: 'Playground',
+      },
+    },
+    title: {
+      name: 'Form message title',
+      if: { arg: 'displayMessage', eq: true },
+      table: {
+        category: 'Playground',
+      },
+    },
+    message: {
+      name: 'Form message text',
+      if: { arg: 'displayMessage', eq: true },
+      table: {
+        category: 'Playground',
+      },
+    },
+    severityLevel: {
+      name: 'Form message severity level',
+      if: { arg: 'displayMessage', eq: true },
+      options: ['warning', 'error'],
+      control: { type: 'inline-radio' },
+      table: {
+        category: 'Playground',
+      },
+    },
+  },
   args: {
     htmlId: 'fksd4kfds',
     label: 'Example label',
     placeholder: 'Placeholder text...',
+    displayMessage: false,
+    message: 'Error',
+    severityLevel: 'warning',
+    title: 'Warning message',
   },
 } as Meta<FormItemComponent>;
 
@@ -24,7 +59,12 @@ const Template: Story<FormItemComponent> = (args: FormItemComponent) => ({
                       [labelId]="labelId"
                       [htmlId]="htmlId">
       <cvi-ng-input [placeholder]="placeholder"
-                    [htmlId]="htmlId">
+                    [htmlId]="htmlId"
+                    cviNgFormMessage
+                    [severityLevel]="severityLevel"
+                    [displayMessage] ="displayMessage"
+                    [message]="message"
+                    [title]="title">
       </cvi-ng-input>
     </cvi-ng-form-item>
   `,
@@ -45,7 +85,12 @@ const WithTextareaTemplate: Story<FormItemComponent> = (
                       [labelId]="labelId"
                       [htmlId]="htmlId">
       <cvi-ng-textarea [placeholder]="placeholder"
-                       [htmlId]="htmlId">
+                       [htmlId]="htmlId"
+                       cviNgFormMessage
+                       [severityLevel]="severityLevel"
+                       [displayMessage] ="displayMessage"
+                       [message]="message"
+                       [title]="title">
       </cvi-ng-textarea>
     </cvi-ng-form-item>
   `,
@@ -84,13 +129,23 @@ const WithInlineFormElementsTemplate: Story<FormItemComponent> = (
                         [labelId]="labelId"
                         [htmlId]="htmlId">
         <cvi-ng-input [placeholder]="placeholder"
-                      [htmlId]="htmlId">
+                      [htmlId]="htmlId"
+                      cviNgFormMessage
+                      [displayMessage]="displayMessage"
+                      [severityLevel]="severityLevel"
+                      [message]="message"
+                      [title]="title">  
         </cvi-ng-input>
       </cvi-ng-form-item>
       <cvi-ng-form-item label="Datepicker"
                         htmlId="datepicker-id">
         <cvi-ng-datepicker placeholder="Pick a date"
-                           htmlId="datepicker-id">
+                           htmlId="datepicker-id"
+                           cviNgFormMessage
+                           [severityLevel]="severityLevel"
+                           [displayMessage] ="displayMessage"
+                           [message]="message"
+                           [title]="title">
         </cvi-ng-datepicker>
       </cvi-ng-form-item>
       <cvi-ng-form-item label="Select"
@@ -99,7 +154,12 @@ const WithInlineFormElementsTemplate: Story<FormItemComponent> = (
                         [ngStyle]="{'width.px': 220}">
         <cvi-ng-select [items]="items"
                        labelId="select-label-id"
-                       htmlId="select-id">
+                       htmlId="select-id"
+                       cviNgFormMessage
+                       [severityLevel]="severityLevel"
+                       [displayMessage] ="displayMessage"
+                       [message]="message"
+                       [title]="title">
         </cvi-ng-select>
       </cvi-ng-form-item>
     </cvi-ng-track>