+++ libs/storybook/CHANGELOG.md
@@ -2,6 +2,33 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [1.12.0](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/compare/storybook-1.11.0...storybook-1.12.0) (2023-09-07)
+
+
+### Features
+
+* **ng:** create form notification component EBS-864  ([#170](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/issues/170)) ([e3a9bc2](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/commits/e3a9bc2fbfd7896cf5a8c1db6801221548a209b4))
+
+
+
+# [1.12.0](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/compare/storybook-1.11.0...storybook-1.12.0) (2023-09-05)
+
+
+### Features
+
+* **ng:** create form notification component EBS-864  ([#170](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/issues/170)) ([e3a9bc2](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/commits/e3a9bc2fbfd7896cf5a8c1db6801221548a209b4))
+
+
+
+# [1.11.0](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/compare/storybook-1.10.2...storybook-1.11.0) (2023-08-29)
+
+
+### Features
+
+* **ng:** box diagram EBS-856 ([#173](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/issues/173)) ([3cd6156](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/commits/3cd615642b04b4839d97b935da8d020b2ff5b92e))
+
+
+
 ## [1.10.2](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/compare/storybook-1.10.1...storybook-1.10.2) (2023-08-02)